+++ lib/yt/models/resource.rb
@@ -8,16 +8,40 @@ module Yt
     class Resource < Base
       attr_reader :auth
 
+    ### ID ###
+
       # @!attribute [r] id
       #   @return [String] the ID that YouTube uses to identify each resource.
       has_one :id
+
+    ### STATUS ###
+
+      has_one :status
+
+      # @return [String] the privacy status of the resource. Valid values are:
+      #   private, public, unlisted.
+      delegate :privacy_status, to: :status
+
+      # @return [Boolean] whether the resource is public.
+      def public?
+        privacy_status == 'public'
+      end
+
+      # @return [Boolean] whether the resource is private.
+      def private?
+        privacy_status == 'private'
+      end
+
+      # @return [Boolean] whether the resource is unlisted.
+      def unlisted?
+        privacy_status == 'unlisted'
+      end
+
 
       has_one :snippet
       delegate :title, :description, :thumbnail_url, :published_at,
         to: :snippet
 
-      has_one :status
-      delegate :privacy_status, :public?, :private?, :unlisted?, to: :status
 
       def initialize(options = {})
         @url = URL.new(options[:url]) if options[:url]